move sync tests into their own file
